create custom hook to use mobile content
